nothing

+++ ds15gl/dsSoldier.h
@@ -7,16 +7,19 @@
 class dsSoldier {
 public:
     dsSoldier();
-    void renderFrame(unsigned int frame_index);
-    void renderSmoothly(int frame1_index, int frame2_index, GLfloat percentage);
-    void renderSmoothly(float progress);
-    void load(const char* person_model_file, const char* person_skin_file, const char* weapon_model_file, const char* weapon_skin_file);
+    void renderFrame(size_t frame_index);
+    void renderSmoothly(size_t frame1_index, size_t frame2_index, GLfloat percentage);
+    void renderSmoothly(GLfloat progress);
+    void load(const char* person_model_file,
+              const char* person_skin_file,
+              const char* weapon_model_file,
+              const char* weapon_skin_file);
     dsTimeManager time_manager;
 private:
     MD2Model person;
     MD2Model weapon;
-    unsigned int frame_beg;
-    unsigned int frame_end;
+    size_t frame_beg;
+    size_t frame_end;
 };
 
 #endif
